# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses (newest first)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses with joined category name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Invalid input or unknown category
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Invalid input or unknown category
        "404":
          description: Not found
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /loa-options:
    get:
      operationId: listLoaOptions
      tags: [loa-options]
      summary: List LOA dropdown options
      parameters:
        - name: category
          in: query
          description: Filter by category (work_type, work_site, job_title)
          schema:
            type: string
            enum: [work_type, work_site, job_title]
      responses:
        "200":
          description: List of options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoaOption"
    post:
      operationId: createLoaOption
      tags: [loa-options]
      summary: Add a new LOA dropdown option
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoaOptionInput"
      responses:
        "201":
          description: Created option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoaOption"
        "409":
          description: Option already exists in this category
  /loa-options/{id}:
    patch:
      operationId: updateLoaOption
      tags: [loa-options]
      summary: Rename a LOA dropdown option
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLoaOptionInput"
      responses:
        "200":
          description: Updated option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoaOption"
        "404":
          description: Not found
        "409":
          description: Another option with this value already exists in this category
    delete:
      operationId: deleteLoaOption
      tags: [loa-options]
      summary: Delete a LOA dropdown option
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoaOption:
      type: object
      required: [id, category, value, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
          enum: [work_type, work_site, job_title]
        value:
          type: string
        createdAt:
          type: string
    LoaOptionInput:
      type: object
      required: [category, value]
      properties:
        category:
          type: string
          enum: [work_type, work_site, job_title]
        value:
          type: string
          minLength: 1
    UpdateLoaOptionInput:
      type: object
      required: [value]
      properties:
        value:
          type: string
          minLength: 1
...
    Expense:
      type: object
      required: [id, categoryId, categoryName, amount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        # numeric(14,2) — sent as a string to preserve precision
        amount:
          type: string
        expenseDate:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ExpenseInput:
      type: object
      required: [categoryId, amount]
      properties:
        categoryId:
          type: integer
        amount:
          type: string
          minLength: 1
        expenseDate:
          type: string
        remarks:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        amount:
          type: string
          minLength: 1
        expenseDate:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
...